+++ cli/tests/test_git_fetch.rs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::io::Write as _;
+
 use testutils::git;
 
 use crate::common::CommandOutput;
@@ -224,6 +226,54 @@ fn test_git_fetch_multiple_remotes() {
       @rem1: ppspxspk 4acd0343 message
     rem2: pzqqpnpo 44c57802 message
       @rem2: pzqqpnpo 44c57802 message
+    [EOF]
+    ");
+}
+
+#[test]
+fn test_git_fetch_with_ignored_refspecs() {
+    let test_env = TestEnvironment::default();
+    test_env.add_config("git.auto-local-bookmark = true");
+    test_env
+        .run_jj_in(".", ["git", "init", "--colocate", "repo"])
+        .success();
+    let source_repo = init_git_remote(&test_env, "origin");
+
+    for branch in ["main", "foo", "foobar", "foobaz", "bar"] {
+        add_commit_to_branch(&source_repo, &format!("refs/heads/{branch}"), branch);
+    }
+
+    let work_dir = test_env.work_dir("repo");
+    std::fs::OpenOptions::new()
+        .append(true)
+        .open(work_dir.root().join("./.git/config"))
+        .expect("failed to open config file")
+        .write_all(
+            br#"
+            [remote "origin"]
+            url = ../origin/.git
+            fetch = +refs/heads/main:refs/remotes/origin/main
+            fetch = +refs/heads/foo*:refs/remotes/origin/baz*
+            fetch = +refs/heads/bar*:refs/tags/bar*
+            fetch = refs/heads/bar
+            "#,
+        )
+        .expect("failed to update config file");
+
+    let output = work_dir.run_jj(["git", "fetch"]).success();
+
+    insta::assert_snapshot!(output.stdout, @r"");
+    insta::assert_snapshot!(output.stderr, @r"
+    Warning: Ignored refspec `refs/heads/bar` from `origin`: fetch-only refspecs are not supported
+    Warning: Ignored refspec `+refs/heads/bar*:refs/tags/bar*` from `origin`: only refs/remotes/ is supported for fetch destinations
+    Warning: Ignored refspec `+refs/heads/foo*:refs/remotes/origin/baz*` from `origin`: renaming is not supported
+    bookmark: main@origin [new] tracked
+    [EOF]
+    ");
+    insta::assert_snapshot!(get_bookmark_output(&work_dir), @r"
+    main: kpukqsrq f803461e main
+      @git: kpukqsrq f803461e main
+      @origin: kpukqsrq f803461e main
     [EOF]
     ");
 }